test: enforce python checks everywhere
- Move all the python style checks into one file
- Run black/isort in separate tests so we get both outputs
- Run black/isort also outside the tests directory
