+++ tests/integration_tests/style/test_python.py
@@ -2,15 +2,52 @@
 # SPDX-License-Identifier: Apache-2.0
 """Tests ensuring codebase style compliance for Python."""
 
-from framework import utils
+import sys
+from subprocess import run
+
+import pytest
 
 
-def test_python_style():
+@pytest.mark.parametrize("formatter", ["black", "isort"])
+def test_python_style(formatter):
     """
-    Test that python code passes style checks.
+    Test that python code passes `formatter`
 
     @type: style
     """
-    # Runs command
-    utils.run_cmd("black --check --diff .")
-    utils.run_cmd("isort --check --diff .")
+    run(
+        f"{formatter} --check --diff . ..",
+        stdout=sys.stdout,
+        stderr=sys.stderr,
+        shell=True,
+        check=True,
+    )
+
+
+def test_python_pylint():
+    """
+    Test that python code passes linter checks.
+
+    @type: build
+    """
+    # List of linter commands that should be executed for each file
+    linter_cmd = (
+        # Pylint
+        "pylint --jobs=0 --persistent=no --score=no "
+        '--output-format=colorized --attr-rgx="[a-z_][a-z0-9_]{1,30}$" '
+        '--argument-rgx="[a-z_][a-z0-9_]{1,35}$" '
+        '--variable-rgx="[a-z_][a-z0-9_]{1,30}$" --disable='
+        "fixme,too-many-instance-attributes,import-error,"
+        "too-many-locals,too-many-arguments,consider-using-f-string,"
+        "consider-using-with,implicit-str-concat,line-too-long,"
+        "broad-except,duplicate-code tests tools .buildkite/*.py"
+    )
+    run(
+        linter_cmd,
+        # we let pytest capture stdout/stderr for us
+        stdout=sys.stdout,
+        stderr=sys.stderr,
+        shell=True,
+        cwd="..",
+        check=True,
+    )
